+++ wxverify/wxverify/__init__.py
@@ -1,3 +1,3 @@
 """Weather Verify Home Assistant add-on package."""
 
-__version__ = "0.8.3"
+__version__ = "0.8.4"